refactor: pass KnowleledgeBaseResponse type to core and import in plguins that use it
