+++ packages/botonic-core/src/models/index.ts
@@ -1,4 +1,5 @@
 export * from './events'
 export * from './hubtype-analytics'
+export * from './knowledge-bases'
 export * from './legacy-types'
 export * from './user'
