+++ packages/botonic-core/src/models/knowledge-bases.ts
@@ -1,0 +1,7 @@
+export interface KnowledgeBaseResponse {
+  inferenceId: string
+  hasKnowledge: boolean
+  isFaithful: boolean
+  chunkIds: string[]
+  answer: string
+}
