+++ packages/botonic-plugin-knowledge-bases/src/types.ts
@@ -1,16 +1,12 @@
+import { KnowledgeBaseResponse } from '@botonic/core'
+
 export interface PluginKnowledgeBaseOptions {
   host: string
   authToken?: string
   timeout?: number
 }
 
-export interface KnowledgeBaseResponse {
-  inferenceId: string
-  hasKnowledge: boolean
-  isFaithful: boolean
-  chunkIds: string[]
-  answer: string
-}
+export type { KnowledgeBaseResponse }
 
 export interface Chunk {
   id: string
